Witt-Trans/sections/ch8/0_texts_mds_english/ragSteps/requirements.txt
pinecone-client[grpc]==3.0.2
openai==1.12.0
python-dotenv==1.0.1 